refactor(agents): split embedded-agent subscribe tool handlers into concept modules
