+++ config/max-lines-baseline.txt
@@ -397,7 +397,6 @@ src/agents/embedded-agent-subscribe.handlers.lifecycle.test.ts
 src/agents/embedded-agent-subscribe.handlers.messages.test.ts
 src/agents/embedded-agent-subscribe.handlers.messages.ts
 src/agents/embedded-agent-subscribe.handlers.tools.test.ts
-src/agents/embedded-agent-subscribe.handlers.tools.ts
 src/agents/embedded-agent-subscribe.subscribe-embedded-agent-session.subscribeembeddedagentsession.test.ts
 src/agents/embedded-agent-subscribe.tools.ts
 src/agents/embedded-agent-subscribe.ts
